changed all double and decimal & database restart

+++ BackEnd/database/migrations/2024_07_23_115501_create_redeem_codes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('redeem_codes', function (Blueprint $table) {
             $table->id();
             $table->string('code');
-            $table->double('value');
+            $table->decimal('value', 19, 4);
             $table->unsignedBigInteger('user_id')->nullable();
             $table->timestamps();
